JIC - Object Values Entered
JIC I wanted to commit these changes and then keep moving so that I dont lose all of the typing I just did lol. I added all of the values into the object for both the card interpretation as well as conversion of that interpretation into number values. Should come in handy!